#94 fix : if 문 누락 추가

+++ com.devsquad10.user/src/main/java/com/devsquad10/user/application/service/UserService.java
@@ -143,7 +143,10 @@ public class UserService {
 	public void deleteUser(UUID id) {
 		User user = (User)userRepository.findByIdAndDeletedAtIsNull(id)
 			.orElseThrow(() -> new IllegalArgumentException("가입되지 않은 사용자입니다."));
-		shippingClient.deleteShippingAgent(id);
+
+		if (user.getRole() == UserRoleEnum.DVL_OFFICER) {
+			shippingClient.deleteShippingAgent(id);
+		}
 
 		user.delete(id);
 		userRepository.save(user);
